requirements.txt
openstacksdk
python-openstackclient